requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.2.3
numpy==1.26.4
plotly==5.17.0
scikit-learn==1.4.2
snowflake-connector-python>=3.12.0
python-dotenv==1.0.0
joblib==1.4.2
pydantic==2.5.0
python-multipart==0.0.6 